refactor(merkl): single source of truth for MERKL_TYPES, derive response keys
Consolidate MerklOpportunityType / MERKL_TYPES into entities/merkl.ts so the
warm-cache plugin, /api/rewards/merkl proxy, and useMerkl composable share
one declaration instead of three drift-prone copies. Replace the proxy's
positional results[0..4] destructuring with iteration over MERKL_TYPES that
derives each response key via Lowercase<MerklOpportunityType> — reordering
the list can no longer silently swap payloads. Drop the duplicate
MerklOpportunityKind union from useMerkl in favor of the shared type.
