+++ entities/merkl.ts
@@ -194,3 +194,23 @@ export interface RewardToken {
   price: number | null
   minimumAmountPerHour: string
 }
+
+export type MerklOpportunityType
+  = | 'EULER'
+    | 'MULTILENDBORROW'
+    | 'ERC20LOGPROCESSOR'
+    | 'EULER_BORROW_FROM_COLLATERAL'
+    | 'EULER_MULTI_BORROW_FROM_COLLATERAL'
+
+// Single source of truth for the proxy response shape and warm-cache fan-out.
+// Lowercase keys (`Lowercase<MerklOpportunityType>`) are derived from these
+// values so reordering this list cannot silently swap response payloads.
+export const MERKL_TYPES: readonly MerklOpportunityType[] = [
+  'EULER',
+  'MULTILENDBORROW',
+  'ERC20LOGPROCESSOR',
+  'EULER_BORROW_FROM_COLLATERAL',
+  'EULER_MULTI_BORROW_FROM_COLLATERAL',
+] as const
+
+export type MerklResponseKey = Lowercase<MerklOpportunityType>
